+++ apps/frontend/src/features/staffs/components/StaffProfileModal.tsx
@@ -1,4 +1,6 @@
 import { Button, Modal } from "@/components/Elements";
+import { useMemo } from "react";
+import { STAFF_EMPTY_IMG_PATH } from "../constants";
 import { StaffProfile } from "../types";
 
 interface StaffProfileProps {
@@ -12,11 +14,21 @@ export const StaffProfileModal = ({
   staffProfile,
   onClose,
 }: StaffProfileProps): JSX.Element => {
+  const isCareerProfile = useMemo(() => {
+    return ![
+      staffProfile.career,
+      staffProfile.license,
+      staffProfile.coachingAchievement,
+      staffProfile.playerHistory,
+      staffProfile.playerAchievement,
+    ].every((item) => !item);
+  }, [staffProfile]);
+
   return (
     <Modal visible={visible} onClose={onClose}>
       <div className="w-[90vw] max-w-[700px] h-[90vh] overflow-scroll p-6">
         <img
-          src={staffProfile.profileImage}
+          src={staffProfile.profileImage || STAFF_EMPTY_IMG_PATH}
           alt={staffProfile.name}
           className="object-cover w-[200px] h-[200px] rounded-[50%] m-auto mb-4"
         />
@@ -30,80 +42,84 @@ export const StaffProfileModal = ({
           {staffProfile.oneWord}
         </p>
 
-        <p className="mb-8 whitespace-pre-line">{staffProfile.description}</p>
+        {staffProfile.description && (
+          <p className="mb-8 whitespace-pre-line">{staffProfile.description}</p>
+        )}
 
-        <div className="bg-gray-200 px-6 py-8 rounded-lg">
-          {/* TODO: リストコンポーネント化 */}
-          {staffProfile.career && (
-            <div className="mb-6">
-              <h3 className="text-xl font-semibold mb-2">経歴</h3>
+        {isCareerProfile && (
+          <div className="bg-gray-200 px-6 py-8 rounded-lg">
+            {/* TODO: リストコンポーネント化 */}
+            {staffProfile.career && (
+              <div className="mb-6">
+                <h3 className="text-xl font-semibold mb-2">経歴</h3>
 
-              <ul>
-                {staffProfile.career.map((item, index) => (
-                  <li className="mb-1" key={index}>
-                    {item}
-                  </li>
-                ))}
-              </ul>
-            </div>
-          )}
+                <ul>
+                  {staffProfile.career.map((item, index) => (
+                    <li className="mb-1" key={index}>
+                      {item}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            )}
 
-          {staffProfile.coachingAchievement && (
-            <div className="mb-6">
-              <h3 className="text-xl font-semibold mb-2">指導実績</h3>
+            {staffProfile.coachingAchievement && (
+              <div className="mb-6">
+                <h3 className="text-xl font-semibold mb-2">指導実績</h3>
 
-              <ul>
-                {staffProfile.coachingAchievement.map((item, index) => (
-                  <li className="mb-1" key={index}>
-                    {item}
-                  </li>
-                ))}
-              </ul>
-            </div>
-          )}
+                <ul>
+                  {staffProfile.coachingAchievement.map((item, index) => (
+                    <li className="mb-1" key={index}>
+                      {item}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            )}
 
-          {staffProfile.license && (
-            <div className="mb-6">
-              <h3 className="text-xl font-semibold mb-2">資格</h3>
+            {staffProfile.license && (
+              <div className="mb-6">
+                <h3 className="text-xl font-semibold mb-2">資格</h3>
 
-              <ul>
-                {staffProfile.license.map((item, index) => (
-                  <li className="mb-1" key={index}>
-                    {item}
-                  </li>
-                ))}
-              </ul>
-            </div>
-          )}
+                <ul>
+                  {staffProfile.license.map((item, index) => (
+                    <li className="mb-1" key={index}>
+                      {item}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            )}
 
-          {staffProfile.playerHistory && (
-            <div className="mb-6">
-              <h3 className="text-xl font-semibold mb-2">選手歴</h3>
+            {staffProfile.playerHistory && (
+              <div className="mb-6">
+                <h3 className="text-xl font-semibold mb-2">選手歴</h3>
 
-              <ul>
-                {staffProfile.playerHistory.map((item, index) => (
-                  <li className="mb-1" key={index}>
-                    {item}
-                  </li>
-                ))}
-              </ul>
-            </div>
-          )}
+                <ul>
+                  {staffProfile.playerHistory.map((item, index) => (
+                    <li className="mb-1" key={index}>
+                      {item}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            )}
 
-          {staffProfile.playerAchievement && (
-            <div className="mb-6">
-              <h3 className="text-xl font-semibold mb-2">選手実績</h3>
+            {staffProfile.playerAchievement && (
+              <div className="mb-6">
+                <h3 className="text-xl font-semibold mb-2">選手実績</h3>
 
-              <ul>
-                {staffProfile.playerAchievement.map((item, index) => (
-                  <li className="mb-1" key={index}>
-                    {item}
-                  </li>
-                ))}
-              </ul>
-            </div>
-          )}
-        </div>
+                <ul>
+                  {staffProfile.playerAchievement.map((item, index) => (
+                    <li className="mb-1" key={index}>
+                      {item}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            )}
+          </div>
+        )}
 
         <div className="w-3/4 m-auto mt-8 px-4">
           {/* NOTE: モーダルを開くタイミングで強制的にボタンまでスクロールしてしまうためtabIndexに負の値を渡す */}
