+++ apps/frontend/src/features/staffs/constants/index.ts
@@ -1,0 +1,1 @@
+export const STAFF_EMPTY_IMG_PATH = "/user-icon.svg";
